+++ LightMessager/DAL/DataBaseEnum.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace LightMessager.DAL
+{
+    public enum DataBaseEnum
+    {
+        Mysql = 1,
+        SqlServer = 2,
+        Oracle = 3
+    }
+}
