+++ LightMessager/Helper/IRabbitMQProducer.cs
@@ -1,0 +1,14 @@
+﻿using LightMessager.Message;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace LightMessager.Helper
+{
+    public interface IRabbitMQProducer
+    {
+        bool Send(BaseMessage message, int delaySend = 0);
+        bool Publish(BaseMessage message, string pattern, int delaySend = 0);
+        bool FanoutPublish(BaseMessage message);
+    }
+}
