+++ LightMessager/Helper/QueueInfo.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace LightMessager.Helper
+{
+    public class QueueInfo
+    {
+        public string Exchange;
+        public string DefaultRouteKey;
+        public string Queue;
+    }
+}
